+++ src/client/mod.rs
@@ -1,13 +1,4 @@
-#[cfg(feature = "rtu")]
-pub mod rtu;
-
-#[cfg(feature = "sync")]
-pub mod sync;
-
-#[cfg(feature = "tcp")]
-pub mod tcp;
-
-use crate::{frame::*, slave::*};
+//! Modbus clients
 
 use std::{
     fmt::Debug,
@@ -16,24 +7,44 @@ use std::{
 
 use async_trait::async_trait;
 
-/// A transport independent asynchronous client trait.
+use crate::{frame::*, slave::*};
+
+#[cfg(feature = "sync")]
+pub mod sync;
+
+#[cfg(feature = "rtu")]
+pub mod rtu;
+
+#[cfg(feature = "tcp")]
+pub mod tcp;
+
+/// Transport independent asynchronous client trait
 #[async_trait]
 pub trait Client: SlaveContext + Send + Debug {
+    /// Invoke a Modbus function
     async fn call(&mut self, request: Request) -> Result<Response, Error>;
 }
 
-/// An asynchronous Modbus reader.
+/// Asynchronous Modbus reader
 #[async_trait]
 pub trait Reader: Client {
+    /// Read multiple coils (0x01)
     async fn read_coils(&mut self, _: Address, _: Quantity) -> Result<Vec<Coil>, Error>;
 
+    /// Read multiple discrete inputs (0x02)
     async fn read_discrete_inputs(&mut self, _: Address, _: Quantity) -> Result<Vec<Coil>, Error>;
 
-    async fn read_input_registers(&mut self, _: Address, _: Quantity) -> Result<Vec<Word>, Error>;
-
+    /// Read multiple holding registers (0x03)
     async fn read_holding_registers(&mut self, _: Address, _: Quantity)
         -> Result<Vec<Word>, Error>;
 
+    /// Read multiple input registers (0x04)
+    async fn read_input_registers(&mut self, _: Address, _: Quantity) -> Result<Vec<Word>, Error>;
+
+    /// Read and write multiple holding registers (0x17)
+    ///
+    /// The write operation is performed before the read unlike
+    /// the name of the operation might suggest!
     async fn read_write_multiple_registers(
         &mut self,
         _: Address,
@@ -43,25 +54,30 @@ pub trait Reader: Client {
     ) -> Result<Vec<Word>, Error>;
 }
 
-/// An asynchronous Modbus writer.
+/// Asynchronous Modbus writer
 #[async_trait]
 pub trait Writer: Client {
+    /// Write a single coil (0x05)
     async fn write_single_coil(&mut self, _: Address, _: Coil) -> Result<(), Error>;
 
-    async fn write_multiple_coils(&mut self, _: Address, _: &[Coil]) -> Result<(), Error>;
-
+    /// Write a single holding register (0x06)
     async fn write_single_register(&mut self, _: Address, _: Word) -> Result<(), Error>;
 
+    /// Write multiple coils (0x0F)
+    async fn write_multiple_coils(&mut self, _: Address, _: &[Coil]) -> Result<(), Error>;
+
+    /// Write multiple holding registers (0x10)
     async fn write_multiple_registers(&mut self, _: Address, _: &[Word]) -> Result<(), Error>;
 }
 
-/// An asynchronous Modbus client context.
+/// Asynchronous Modbus client context
 #[derive(Debug)]
 pub struct Context {
     client: Box<dyn Client>,
 }
 
 impl Context {
+    /// Disconnect the client
     pub async fn disconnect(&mut self) -> Result<(), Error> {
         // Disconnecting is expected to fail!
         let res = self.client.call(Request::Disconnect).await;
@@ -110,8 +126,8 @@ impl Reader for Context {
         let rsp = self.client.call(Request::ReadCoils(addr, cnt)).await?;
 
         if let Response::ReadCoils(mut coils) = rsp {
-            debug_assert!(coils.len() >= cnt as usize);
-            coils.truncate(cnt as usize);
+            debug_assert!(coils.len() >= cnt.into());
+            coils.truncate(cnt.into());
             Ok(coils)
         } else {
             Err(Error::new(ErrorKind::InvalidData, "unexpected response"))
@@ -129,8 +145,8 @@ impl Reader for Context {
             .await?;
 
         if let Response::ReadDiscreteInputs(mut coils) = rsp {
-            debug_assert!(coils.len() >= cnt as usize);
-            coils.truncate(cnt as usize);
+            debug_assert!(coils.len() >= cnt.into());
+            coils.truncate(cnt.into());
             Ok(coils)
         } else {
             Err(Error::new(ErrorKind::InvalidData, "unexpected response"))
@@ -148,7 +164,7 @@ impl Reader for Context {
             .await?;
 
         if let Response::ReadInputRegisters(rsp) = rsp {
-            if rsp.len() != cnt as usize {
+            if rsp.len() != cnt.into() {
                 return Err(Error::new(ErrorKind::InvalidData, "invalid response"));
             }
             Ok(rsp)
@@ -168,7 +184,7 @@ impl Reader for Context {
             .await?;
 
         if let Response::ReadHoldingRegisters(rsp) = rsp {
-            if rsp.len() != cnt as usize {
+            if rsp.len() != cnt.into() {
                 return Err(Error::new(ErrorKind::InvalidData, "invalid response"));
             }
             Ok(rsp)
@@ -195,7 +211,7 @@ impl Reader for Context {
             .await?;
 
         if let Response::ReadWriteMultipleRegisters(rsp) = rsp {
-            if rsp.len() != read_cnt as usize {
+            if rsp.len() != read_cnt.into() {
                 return Err(Error::new(ErrorKind::InvalidData, "invalid response"));
             }
             Ok(rsp)
@@ -235,7 +251,7 @@ impl Writer for Context {
             .await?;
 
         if let Response::WriteMultipleCoils(rsp_addr, rsp_cnt) = rsp {
-            if rsp_addr != addr || rsp_cnt as usize != cnt {
+            if rsp_addr != addr || usize::from(rsp_cnt) != cnt {
                 return Err(Error::new(ErrorKind::InvalidData, "invalid response"));
             }
             Ok(())
@@ -276,7 +292,7 @@ impl Writer for Context {
             .await?;
 
         if let Response::WriteMultipleRegisters(rsp_addr, rsp_cnt) = rsp {
-            if rsp_addr != addr || rsp_cnt as usize != cnt {
+            if rsp_addr != addr || usize::from(rsp_cnt) != cnt {
                 return Err(Error::new(ErrorKind::InvalidData, "invalid response"));
             }
             Ok(())