+++ src/codec/mod.rs
@@ -7,10 +7,28 @@ pub mod tcp;
 
 use crate::frame::*;
 
-use byteorder::{BigEndian, ReadBytesExt};
+use byteorder::{BigEndian, ReadBytesExt as _};
 use bytes::{BufMut, Bytes, BytesMut};
 use std::convert::TryFrom;
 use std::io::{self, Cursor, Error, ErrorKind};
+
+#[allow(clippy::cast_possible_truncation)]
+fn u16_len(len: usize) -> u16 {
+    // This type conversion should always be safe, because either
+    // the caller is responsible to pass a valid usize or the
+    // possible values are limited by the protocol.
+    debug_assert!(len <= u16::MAX.into());
+    len as u16
+}
+
+#[allow(clippy::cast_possible_truncation)]
+fn u8_len(len: usize) -> u8 {
+    // This type conversion should always be safe, because either
+    // the caller is responsible to pass a valid usize or the
+    // possible values are limited by the protocol.
+    debug_assert!(len <= u8::MAX.into());
+    len as u8
+}
 
 impl From<Request> for Bytes {
     fn from(req: Request) -> Bytes {
@@ -33,9 +51,9 @@ impl From<Request> for Bytes {
             WriteMultipleCoils(address, coils) => {
                 data.put_u16(address);
                 let len = coils.len();
-                data.put_u16(len as u16);
+                data.put_u16(u16_len(len));
                 let packed_coils = pack_coils(&coils);
-                data.put_u8(packed_coils.len() as u8);
+                data.put_u8(u8_len(packed_coils.len()));
                 for b in packed_coils {
                     data.put_u8(b);
                 }
@@ -47,8 +65,8 @@ impl From<Request> for Bytes {
             WriteMultipleRegisters(address, words) => {
                 data.put_u16(address);
                 let len = words.len();
-                data.put_u16(len as u16);
-                data.put_u8((len as u8) * 2);
+                data.put_u16(u16_len(len));
+                data.put_u8(u8_len(len * 2));
                 for w in words {
                     data.put_u16(w);
                 }
@@ -58,8 +76,8 @@ impl From<Request> for Bytes {
                 data.put_u16(quantity);
                 data.put_u16(write_address);
                 let n = words.len();
-                data.put_u16(n as u16);
-                data.put_u8(n as u8 * 2);
+                data.put_u16(u16_len(n));
+                data.put_u8(u8_len(n * 2));
                 for w in words {
                     data.put_u16(w);
                 }
@@ -94,7 +112,7 @@ impl From<Response> for Bytes {
         match rsp {
             ReadCoils(coils) | ReadDiscreteInputs(coils) => {
                 let packed_coils = pack_coils(&coils);
-                data.put_u8(packed_coils.len() as u8);
+                data.put_u8(u8_len(packed_coils.len()));
                 for b in packed_coils {
                     data.put_u8(b);
                 }
@@ -102,7 +120,7 @@ impl From<Response> for Bytes {
             ReadInputRegisters(registers)
             | ReadHoldingRegisters(registers)
             | ReadWriteMultipleRegisters(registers) => {
-                data.put_u8((registers.len() * 2) as u8);
+                data.put_u8(u8_len(registers.len() * 2));
                 for r in registers {
                     data.put_u16(r);
                 }
@@ -135,7 +153,7 @@ impl From<ExceptionResponse> for Bytes {
         let mut data = BytesMut::with_capacity(2);
         debug_assert!(ex.function < 0x80);
         data.put_u8(ex.function + 0x80);
-        data.put_u8(ex.exception as u8);
+        data.put_u8(ex.exception.into());
         data.freeze()
     }
 }
@@ -165,7 +183,7 @@ impl TryFrom<Bytes> for Request {
                 let address = rdr.read_u16::<BigEndian>()?;
                 let quantity = rdr.read_u16::<BigEndian>()?;
                 let byte_count = rdr.read_u8()?;
-                if bytes.len() < (6 + byte_count as usize) {
+                if bytes.len() < 6 + usize::from(byte_count) {
                     return Err(Error::new(ErrorKind::InvalidData, "Invalid byte count"));
                 }
                 let x = &bytes[6..];
@@ -180,11 +198,11 @@ impl TryFrom<Bytes> for Request {
             0x10 => {
                 let address = rdr.read_u16::<BigEndian>()?;
                 let quantity = rdr.read_u16::<BigEndian>()?;
-                let byte_count = rdr.read_u8()? as usize;
-                if bytes.len() < (6 + byte_count as usize) {
+                let byte_count = rdr.read_u8()?;
+                if bytes.len() < 6 + usize::from(byte_count) {
                     return Err(Error::new(ErrorKind::InvalidData, "Invalid byte count"));
                 }
-                let mut data = vec![];
+                let mut data = Vec::with_capacity(quantity.into());
                 for _ in 0..quantity {
                     data.push(rdr.read_u16::<BigEndian>()?);
                 }
@@ -195,11 +213,11 @@ impl TryFrom<Bytes> for Request {
                 let read_quantity = rdr.read_u16::<BigEndian>()?;
                 let write_address = rdr.read_u16::<BigEndian>()?;
                 let write_quantity = rdr.read_u16::<BigEndian>()?;
-                let write_count = rdr.read_u8()? as usize;
-                let mut data = vec![];
-                if bytes.len() < (10 + write_count as usize) {
+                let write_count = rdr.read_u8()?;
+                if bytes.len() < 10 + usize::from(write_count) {
                     return Err(Error::new(ErrorKind::InvalidData, "Invalid byte count"));
                 }
+                let mut data = Vec::with_capacity(write_quantity.into());
                 for _ in 0..write_quantity {
                     data.push(rdr.read_u16::<BigEndian>()?);
                 }
@@ -258,7 +276,7 @@ impl TryFrom<Bytes> for Response {
             0x04 => {
                 let byte_count = rdr.read_u8()?;
                 let quantity = byte_count / 2;
-                let mut data = vec![];
+                let mut data = Vec::with_capacity(quantity.into());
                 for _ in 0..quantity {
                     data.push(rdr.read_u16::<BigEndian>()?);
                 }
@@ -267,7 +285,7 @@ impl TryFrom<Bytes> for Response {
             0x03 => {
                 let byte_count = rdr.read_u8()?;
                 let quantity = byte_count / 2;
-                let mut data = vec![];
+                let mut data = Vec::with_capacity(quantity.into());
                 for _ in 0..quantity {
                     data.push(rdr.read_u16::<BigEndian>()?);
                 }
@@ -281,7 +299,7 @@ impl TryFrom<Bytes> for Response {
             0x17 => {
                 let byte_count = rdr.read_u8()?;
                 let quantity = byte_count / 2;
-                let mut data = vec![];
+                let mut data = Vec::with_capacity(quantity.into());
                 for _ in 0..quantity {
                     data.push(rdr.read_u16::<BigEndian>()?);
                 }
@@ -376,15 +394,15 @@ fn pack_coils(coils: &[Coil]) -> Vec<u8> {
     let mut res = vec![0; packed_size];
     for (i, b) in coils.iter().enumerate() {
         let v = if *b { 0b1 } else { 0b0 };
-        res[(i / 8) as usize] |= v << (i % 8);
+        res[i / 8] |= v << (i % 8);
     }
     res
 }
 
 fn unpack_coils(bytes: &[u8], count: u16) -> Vec<Coil> {
-    let mut res = Vec::with_capacity(count as usize);
-    for i in 0..count {
-        res.push((bytes[(i / 8u16) as usize] >> (i % 8)) & 0b1 > 0);
+    let mut res = Vec::with_capacity(count.into());
+    for i in 0usize..count.into() {
+        res.push((bytes[i / 8] >> (i % 8)) & 0b1 > 0);
     }
     res
 }
@@ -980,7 +998,7 @@ mod tests {
         fn read_coils_max_quantity() {
             let quantity = 2000;
             let byte_count = quantity / 8;
-            let mut raw: Vec<u8> = vec![1, byte_count as u8];
+            let mut raw: Vec<u8> = vec![1, u8_len(byte_count)];
             let mut values: Vec<u8> = (0..byte_count).map(|_| 0b_1111_1111).collect();
             raw.append(&mut values);
             let bytes = Bytes::from(raw);
@@ -1004,7 +1022,7 @@ mod tests {
         fn read_discrete_inputs_max_quantity() {
             let quantity = 2000;
             let byte_count = quantity / 8;
-            let mut raw: Vec<u8> = vec![2, byte_count as u8];
+            let mut raw: Vec<u8> = vec![2, u8_len(byte_count)];
             let mut values: Vec<u8> = (0..byte_count).map(|_| 0b_1111_1111).collect();
             raw.append(&mut values);
             let bytes = Bytes::from(raw);